pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ehr-fm"
version = "0.1.0"
description = "EHR : Foundational Model"
readme = "README.md"
authors = [
    { name = "Sudhanva", email = "athreya.sud@gmail.com" },
]
maintainers = [
    { name = "Sudhanva", email = "athreya.sud@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "torch",
    "transformers",
    "polars",
    "hydra-core",
    "hydra-joblib-launcher",
    "loguru",
    "safetensors",
    "tqdm",
    "wandb",
    "mlflow",
    "MEDS_transforms~=0.1.0",
    "pyarrow",
    "numpy",
    "pytorch-lightning",
    "tensorboard",
    "einops",
    "python-dotenv",
    "boto3",
]

[project.optional-dependencies]
dev = ["pre-commit<4"]
tests = ["pytest"]
jupyter = ["jupyter", "seaborn", "scikit-learn", "matplotlib", "pandas", "altair", "pylatex", "vl-convert-python"]
eval = ["scikit-learn", "psutil", "pyyaml"]

#[tool.setuptools.package-data]
#ethos = ["tokenize/maps/*.csv.gz"]